+++ tsfile/src/main/java/org/apache/iotdb/tsfile/read/common/block/column/TimeColumn.java
@@ -27,7 +27,7 @@ import static org.apache.iotdb.tsfile.read.common.block.column.ColumnUtil.checkV
 
 public class TimeColumn implements Column {
 
-  private static final int INSTANCE_SIZE = ClassLayout.parseClass(LongColumn.class).instanceSize();
+  private static final long INSTANCE_SIZE = ClassLayout.parseClass(LongColumn.class).instanceSize();
   public static final int SIZE_IN_BYTES_PER_POSITION = Long.BYTES;
 
   private final int arrayOffset;
@@ -145,6 +145,6 @@ public class TimeColumn implements Column {
 
   @Override
   public int getInstanceSize() {
-    return INSTANCE_SIZE;
+    return (int) INSTANCE_SIZE;
   }
 }
